задание 3 / сделали рабочую навигацию по фрагментам
