+++ app/src/main/java/com/example/nasacustomapp/view/WikiFragment.kt
@@ -1,57 +1,57 @@
 package com.example.nasacustomapp.view
 
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.example.nasacustomapp.R
+import com.example.nasacustomapp.databinding.FragmentPhotoOfTheDayBinding
+import com.example.nasacustomapp.databinding.FragmentWikiBinding
+import com.example.nasacustomapp.utils.WIKI_PARSE_URL
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
 
-/**
- * A simple [Fragment] subclass.
- * Use the [WikiFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
 class WikiFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
-        }
-    }
+    private var _binding: FragmentWikiBinding? = null
+    private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_wiki, container, false)
+        _binding = FragmentWikiBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
     companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment WikiFragment.
-         */
-        // TODO: Rename and change types and number of parameters
-        @JvmStatic
+
         fun newInstance() =
             WikiFragment().apply {
 
             }
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        setOnWikiSearchClickListener()
+    }
+    private fun setOnWikiSearchClickListener() {
+
+        binding.inputLayout.setEndIconOnClickListener {
+            startActivity(Intent(Intent.ACTION_VIEW).apply {
+                data = Uri.parse("$WIKI_PARSE_URL${binding.inputEditText.text.toString()}")
+            })
+        }
+
+
+
+
+    }
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding=null
     }
 }
