+++ app/src/main/res/layout/fragment_photo_of_the_day.xml
@@ -16,31 +16,7 @@
             android:paddingStart="10dp"
             android:paddingEnd="10dp">
 
-            <com.google.android.material.textfield.TextInputLayout
-                android:id="@+id/input_layout"
-                style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:hint="@string/search_wiki"
 
-                app:counterEnabled="true"
-                app:counterMaxLength="10"
-
-                app:endIconCheckable="true"
-                app:endIconDrawable="@drawable/ic_wikipedia"
-                app:endIconMode="custom"
-
-                app:helperText="@string/wiki_helper_text"
-                app:layout_constraintBottom_toTopOf="@id/title_text"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent">
-
-                <com.google.android.material.textfield.TextInputEditText
-                    android:id="@+id/input_edit_text"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:inputType="text" />
-            </com.google.android.material.textfield.TextInputLayout>
 
             <TextView
                 android:layout_width="wrap_content"
@@ -48,7 +24,7 @@
                 android:id="@+id/title_text"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="@id/input_layout"
+                app:layout_constraintTop_toTopOf="parent"
                 tools:text="Title"
                 />
